refactor: move ranking list to his own service layer
