+++ src/views/Profile/index.tsx
@@ -8,7 +8,7 @@ import { useRanking } from '~/hooks/useRanking'
 
 export const Profile = () => {
   const { userData } = useAuth()
-  const { userRankingPosition } = useRanking(userData.id)
+  const { userRankingPosition, isLoadingRanking } = useRanking(userData.id)
 
   return (
     <Background>
@@ -24,7 +24,10 @@ export const Profile = () => {
       <S.CardsContainer>
         <DataCard title={userData.points} description="pontos" />
         <DataCard title={userData.time} description="minutos" />
-        <DataCard title={`#${userRankingPosition}`} description="no ranking" />
+        <DataCard
+          title={`#${isLoadingRanking ? '...' : userRankingPosition}`}
+          description="no ranking"
+        />
       </S.CardsContainer>
     </Background>
   )
